requirements.txt
# Dev environment
pip
autopep8
python-dotenv

# App
google-genai # Client library for the Gemini API
Pillow # PIL image-loading library
flask